requirements.txt
blinker==1.6.2
boto3==1.28.3
botocore==1.31.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
PyMySQL==1.1.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
